2 alteração em tipo de retorno; 6 alterações em parametros de função; 2 alteração em tipo de varivel; 6 novos metodos;

+++ src/main/java/org/testes/Circulo.java
@@ -1,9 +1,9 @@
 package org.testes;
 
 public class Circulo extends Forma{
-    int raio;
+    double raio;
 
-    public Circulo(int raio) {
+    public Circulo(double raio) {
         super();
         this.raio = raio;
     }
